# Public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IXXXX": {
				"properties": {
					"ZZZZ": {
						"type": "string"
					},
					"WWWW": {
						"type": "number",
						"format": "double"
					},
					"TTTT": {
						"type": "boolean"
					}
				},
				"required": [
					"ZZZZ",
					"WWWW"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponseOK_IXXXX_": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/IXXXX"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IYYYY": {
				"properties": {
					"KKKK": {
						"type": "string"
					},
					"TTTT": {
						"type": "number",
						"format": "double"
					},
					"JJJJ": {
						"type": "boolean"
					}
				},
				"required": [
					"KKKK",
					"TTTT"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponseOK_IYYYY_": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/IYYYY"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "nodejs",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/XXXX/test/{id}": {
			"get": {
				"operationId": "GetOneTestXXXX",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponseOK_IXXXX_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/XXXX/detail/{id}": {
			"get": {
				"operationId": "GetOneXXXX",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponseOK_IXXXX_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/XXXX/create": {
			"post": {
				"operationId": "CreateOneXXXX",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponseOK_IXXXX_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IXXXX"
							}
						}
					}
				}
			}
		},
		"/YYYY/test/{id}": {
			"get": {
				"operationId": "GetOneTestYYYY",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/YYYY/detail/{id}": {
			"get": {
				"operationId": "GetOneYYYY",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponseOK_IYYYY_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/YYYY/create": {
			"post": {
				"operationId": "CreateOneYYYY",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponseOK_IYYYY_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IYYYY"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}